refactor: migrated population data management commands
